// desktop/src/open-api/model/pagedRoleRequestCommand.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SortDirection } from './sortDirection';
import { RoleFilter } from './roleFilter';


export interface PagedRoleRequestCommand { 
    /**
     * Page number
     */
    page: number;
    /**
     * Number of records per page
     */
    pageSize: number;
    /**
     * field to order on
     */
    orderBy?: string;
    sortDirection?: SortDirection;
    filter?: RoleFilter;
}
export namespace PagedRoleRequestCommand {
}